Add delete functionality and more small edits
This commit installed headless UI and lucude react for the new delete modal.
Made a new removeListing function to update the state in the app component
and added type prop to Button component. Changed the type for listing ids to be string
OR number. Converted the listing form to react hook form as well.