Continue redesign of navbar 2.0
add divs and main element in App component to stick footer at bottom. Removed
the hamburger menu and kept it as a avatar circle as well as added logo
next to the name. Moved nav items for smaller devices into the avatar menu.
Made the width of the search bar to be w-64 and not full. Created a custom
breakpoint called mdlg for 875px and added new logo assets.